Cleanly separate Gtk and SVG implementations
This allows one to precompile and use it with either Gtk or
IJulia.
